Extract NativeResourceWriter
