+++ src/Compilers/Core/Portable/CodeAnalysis.csproj
@@ -72,6 +72,7 @@
     <Compile Include="InternalUtilities\SetWithInsertionOrder.cs" />
     <Compile Include="InternalUtilities\StackGuard.cs" />
     <Compile Include="InternalUtilities\StreamExtensions.cs" />
+    <Compile Include="PEWriter\NativeResourceWriter.cs" />
     <Compile Include="System.Reflection.Metadata\Ecma335\CodedIndex.cs" />
     <Compile Include="System.Reflection.Metadata\Ecma335\EncFuncCode.cs" />
     <Compile Include="System.Reflection.Metadata\Ecma335\HeapSizeFlag.cs" />
